Prints error message whenever a file does not exist

+++ cat.rb
@@ -1,8 +1,12 @@
 def cat(files)
   files.each do |file|
-    input_file = File.open(file, 'r')
-    opened_file = input_file.read
-    puts opened_file
+    if File.exist?(file)
+      input_file = File.open(file, 'r')
+      opened_file = input_file.read
+      puts opened_file
+    else
+      raise ArgumentError, "This file doesn't exist"
+    end
   end
 end
